feat: :recycle: Merge conf and init in a single keycloak file
