+++ apps/client/src/api/xhr-client.spec.js
@@ -1,7 +1,7 @@
 import { vi, describe, it, expect } from 'vitest'
 import * as xhrClient from './xhr-client.js'
 
-vi.mock('@/utils/keycloak/init.js', () => ({
+vi.mock('@/utils/keycloak/keycloak.js', () => ({
   getKeycloak: () => ({
     token: 'token',
     updateToken: vi.fn(),
