--- apps/client/src/utils/keycloak/config.js
@@ -1,16 +1,0 @@
-import {
-  keycloakProtocol,
-  keycloakDomain,
-  keycloakClientId,
-  keycloakRealm,
-  keycloakRedirectUri,
-} from '../env.js'
-
-export const keycloakConf = {
-  url: `${keycloakProtocol}://${keycloakDomain}`,
-  realm: keycloakRealm,
-  clientId: keycloakClientId,
-  onLoad: 'check-sso',
-  flow: 'hybrid',
-  redirectUri: keycloakRedirectUri,
-}
