+++ apps/client/src/utils/keycloak/keycloak.js
@@ -1,5 +1,21 @@
 import Keycloak from 'keycloak-js'
-import { keycloakConf } from './config.js'
+
+import {
+  keycloakProtocol,
+  keycloakDomain,
+  keycloakClientId,
+  keycloakRealm,
+  keycloakRedirectUri,
+} from '../env.js'
+
+export const keycloakConf = {
+  url: `${keycloakProtocol}://${keycloakDomain}`,
+  realm: keycloakRealm,
+  clientId: keycloakClientId,
+  onLoad: 'check-sso',
+  flow: 'hybrid',
+  redirectUri: keycloakRedirectUri,
+}
 
 let keycloak
 
@@ -26,8 +42,8 @@ export const getUserProfile = async () => {
 }
 
 export const keycloakInit = async () => {
+  keycloakConf.redirectUri = location
   try {
-    keycloakConf.redirectUri = location
     const { onLoad, redirectUri, flow } = keycloakConf
     const keycloak = getKeycloak()
     await keycloak.init({ onLoad, flow, redirectUri })
